외부설정 변경 및 브랜드 목록에 미리보기 기능 추가

+++ src/main/resources/static/js/common.js
@@ -32,7 +32,7 @@ const CommonUtils = {
         return targetValue.toString().replace(/\B(?=(\d{3})+(?!\d))/g, ",");
     },
     removeComma(targetValue) {
-        return parseInt(targetValue.replace(/,/g,""));
+        return parseFloat(targetValue.replace(/,/g,""));
     },
     keyEventOnlyNumber(obj, blnComma) {
         if (blnComma) obj.value = obj.value.replace(/[^0-9\.]/g,'').replace(/\B(?=(\d{3})+(?!\d))/g, ",");
